General(:zap:): Abstract header and card components.
