+++ src/pages/ProjectDetails/components/Card/index.ts
